Feature: toggable modules.
Modules are now loaded according to whether the relative setting is
enabled. This makes them toggable from within the application.
Currently requires restart.

In order to make space for a page that allows to toggle the widgets,
the settings dialog has been converted to a tabbed setup, separating
aesthetic modifications from widget toggling.